requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
python-telegram-bot==13.7  # Stick with 13.7 for now; upgrade later if needed
schedule>=1.2.0
pytz>=2023.3
python-dotenv>=1.0.0
streamlit>=1.30.0
